+++ src/record.h
@@ -4,7 +4,7 @@
 typedef struct
 {
 	std::map<std::string, std::string> header;
-	std::vector<int> records;
+	std::vector<float> records;
 	size_t number_samples;
 } RECORD;
